refactor(virtual-keys): split type tests for GH727
